Some tilts made easier
up tilt in particular has been made easier to do, when the stick i
below the middle of the stick, then it will work. if the stick has been
above it for 5 frames or more then you can tilt as well. same will go
for down tilt when it is implemented